requirements.txt
absl-py==0.8.1
astor==0.8.0
gast==0.3.2
grpcio==1.24.1
imageio==2.6.1
Markdown==3.1.1
numpy==1.14.5
Pillow==6.2.0
pkg-resources==0.0.0
protobuf==3.10.0
six==1.12.0
tensorboard==1.9.0
tensorflow-gpu==1.9.0
termcolor==1.1.0
Werkzeug==0.16.0
